requirements.txt
Django==4.2.19
djangorestframework==3.15.2
funasr==1.1.12
pycryptodome==3.21.0
PyJWT==2.9.0
PyJWT==2.10.1
Requests==2.32.3
